# swagger.json
{
    "openapi": "3.0.0",
...
    "info": {
        "version": "1.0.0",
        "title": "API de Cortes de Luz",
        "description": "API utilizada para gestionar los datos de los domicilios, empresas y cortes de luz registrados"
    },
    "servers": [
        {
            "url": "http://localhost:2414"
        }
    ],
...
    "paths": {
...
        "/domicilios": {
            "get": {
                "tags": [
                    "Domicilios"
                ],
                "summary": "Obtener una lista de todos los domicilios",
                "description": "Obtiene una lista de domicilios con sus respectivos datos",
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "direccion": {
                                                "type": "string"
                                            },
                                            "barrio": {
                                                "type": "string"
                                            },
                                            "consumo": {
                                                "type": "number"
                                            },
                                            "dueño": {
                                                "type": "string"
                                            },
                                            "nombre_empresa": {
                                                "type": "string"
                                            },
                                            "ids_cortes": {
                                                "type": "array",
                                                "items": {
                                                    "type": "object",
                                                    "properties": {
                                                        "id_corte": {
                                                            "type": "number"
                                                        }
                                                    }
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Domicilios"
                ],
                "summary": "Agregar un objeto de tipo domicilio a la lista",
                "description": "Agrega un domicilio a la lista de domicilios",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "barrio": {
                                        "type": "string",
                                        "example": "Ejemplo"
                                    },
                                    "direccion": {
                                        "type": "string",
                                        "example": "Avenida"
                                    },
                                    "consumo": {
                                        "type": "number",
                                        "example": 0
                                    },
                                    "dueño": {
                                        "type": "string",
                                        "example": "Ejemplo"
                                    },
                                    "nombre_empresa": {
                                        "type": "string"
                                    },
                                    "ids_cortes": {
                                        "type": "array",
                                        "items": {
                                            "type": "object",
                                            "properties": {
                                                "id_corte": {
                                                    "type": "number"
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "direccion": {
                                                "type": "string"
                                            },
                                            "barrio": {
                                                "type": "string"
                                            },
                                            "consumo": {
                                                "type": "number"
                                            },
                                            "dueño": {
                                                "type": "string"
                                            },
                                            "nombre_empresa": {
                                                "type": "string"
                                            },
                                            "ids_cortes": {
                                                "type": "array",
                                                "items": {
                                                    "type": "object",
                                                    "properties": {
                                                        "id_corte": {
                                                            "type": "number"
                                                        }
                                                    }
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/domicilios/{direccion}": {
            "get": {
                "tags": [
                    "Domicilios"
                ],
                "summary": "Obtener un domicilio buscandolo por su ID",
                "description": "Obtiene un domicilio con sus respectivos datos",
                "parameters": [
                    {
                        "name": "direccion",
                        "in": "path",
                        "schema": {
                            "type": "string"
                        },
                        "required": "true"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "direccion": {
                                                "type": "string"
                                            },
                                            "barrio": {
                                                "type": "string"
                                            },
                                            "consumo": {
                                                "type": "number"
                                            },
                                            "dueño": {
                                                "type": "string"
                                            },
                                            "nombre_empresa": {
                                                "type": "string"
                                            },
                                            "ids_cortes": {
                                                "type": "array",
                                                "items": {
                                                    "id_corte": {
                                                        "type": "number"
                                                    }
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Domicilios"
                ],
                "summary": "Borrar un domicilio buscandolo por su ID",
                "description": "Borra un domicilio",
                "parameters": [
                    {
                        "name": "direccion",
                        "in": "path",
                        "schema": {
                            "type": "string"
                        },
                        "required": "true"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "direccion": {
                                                "type": "string"
                                            },
                                            "barrio": {
                                                "type": "string"
                                            },
                                            "consumo": {
                                                "type": "number"
                                            },
                                            "dueño": {
                                                "type": "string"
                                            },
                                            "nombre_empresa": {
                                                "type": "string"
                                            },
                                            "ids_cortes": {
                                                "type": "array",
                                                "items": {
                                                    "type": "object",
                                                    "properties": {
                                                        "id_corte": {
                                                            "type": "number"
                                                        }
                                                    }
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Domicilios"
                ],
                "summary": "Actualizar todos los datos de un domicilio",
                "description": "Actualiza los valores de un domicilio",
                "parameters": [
                    {
                        "name": "direccion",
                        "in": "path",
                        "schema": {
                            "type": "string"
                        },
                        "required": "true"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "barrio": {
                                        "type": "string",
                                        "example": "Ejemplo"
                                    },
                                    "direccion": {
                                        "type": "string",
                                        "example": "Avenida"
                                    },
                                    "consumo": {
                                        "type": "number",
                                        "example": 0
                                    },
                                    "dueño": {
                                        "type": "string",
                                        "example": "Ejemplo"
                                    },
                                    "nombre_empresa": {
                                        "type": "string",
                                        "example": "Ejemplo"
                                    },
                                    "ids_cortes": {
                                        "type": "array",
                                        "items": {
                                            "type": "object",
                                            "properties": {
                                                "id_corte": {
                                                    "type": "number"
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "direccion": {
                                                "type": "string"
                                            },
                                            "barrio": {
                                                "type": "string"
                                            },
                                            "consumo": {
                                                "type": "number"
                                            },
                                            "dueño": {
                                                "type": "string"
                                            },
                                            "nombre_empresa": {
                                                "type": "string",
                                                "example": "Ejemplo"
                                            },
                                            "ids_cortes": {
                                                "type": "array",
                                                "items": {
                                                    "type": "object",
                                                    "properties": {
                                                        "id_corte": {
                                                            "type": "number"
                                                        }
                                                    }
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Domicilios"
                ],
                "summary": "Actualizar algunos datos de un domicilio",
                "description": "Actualiza ciertos valores de un domicilio",
                "parameters": [
                    {
                        "name": "direccion",
                        "in": "path",
                        "schema": {
                            "type": "string"
                        },
                        "required": "true"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "barrio": {
                                        "type": "string",
                                        "example": "Ejemplo"
                                    },
                                    "direccion": {
                                        "type": "string",
                                        "example": "Avenida"
                                    },
                                    "consumo": {
                                        "type": "number",
                                        "example": 0
                                    },
                                    "dueño": {
                                        "type": "string",
                                        "example": "Ejemplo"
                                    },
                                    "nombre_empresa": {
                                        "type": "string",
                                        "example": "Ejemplo"
                                    },
                                    "ids_cortes": {
                                        "type": "array",
                                        "items": {
                                            "type": "object",
                                            "properties": {
                                                "id_corte": {
                                                    "type": "number"
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "direccion": {
                                                "type": "string"
                                            },
                                            "barrio": {
                                                "type": "string"
                                            },
                                            "consumo": {
                                                "type": "number"
                                            },
                                            "dueño": {
                                                "type": "string"
                                            },
                                            "nombre_empresa": {
                                                "type": "string",
                                                "example": "Ejemplo"
                                            },
                                            "ids_cortes": {
                                                "type": "array",
                                                "items": {
                                                    "type": "object",
                                                    "properties": {
                                                        "id_corte": {
                                                            "type": "number"
                                                        }
                                                    }
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/domicilios/{barrio}/cantidad": {
            "get": {
                "tags": [
                    "Metodos personalizados"
                ],
                "summary": "Busca la cantidad de domicilios por barrio",
                "description": "Cuenta la cantidad de domicilios por barrio verificando las igualdades",
                "parameters": [
                    {
                        "name": "barrio",
                        "in": "path",
                        "schema": {
                            "type": "string"
                        },
                        "required": "true"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "duracion": {
                                                "type": "string",
                                                "example": "Ejemplo"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}